itda-work/skills/slide-ai/requirements.txt
google-genai>=1.60,<2
python-pptx>=1.0,<2
Pillow>=10.0,<12
PyPDF2>=3.0,<4
